dashboard/requirements.txt
psycopg2--binary
python--dotenv
requests
streamlit
altair
pandas